cityId. Long -> Int

+++ app/src/main/java/com/example/seeweather/data/model/CurrentWeatherEntity.kt
@@ -1,7 +1,7 @@
 package com.example.seeweather.data.model
 
 class CurrentWeatherEntity(
-	var cityId: Long = 0L,
+	var cityId: Int = 0,
 	var tempC: Float = 0f,
 	var tempF: Float = 0f,
 	var icon: String = "",
